Created utils.js for useful helper functions. Renamed references of Game ID to Match ID.
